move let, vars, globals tests out to their own teest module (test_basic is getting big)
